move button tokens
