+++ packages/react-button/src/components/Button/Button.types.ts
@@ -1,8 +1,9 @@
 import * as React from 'react';
 import { BaseSlots, SlotProps } from '@fluentui/react-compose';
 import { ComponentProps, ShorthandProps } from '@fluentui/react-compose/lib/next/index';
-import { ColorTokenSet } from '@fluentui/react-theme-provider';
-import { RecursivePartial } from '../../utils/tempTypes';
+import { ButtonTokenSet, RecursivePartial } from '@fluentui/themes';
+
+export { ButtonTokenSet, RecursivePartial } from '@fluentui/themes';
 
 export type SizeValue = 'smallest' | 'smaller' | 'small' | 'medium' | 'large' | 'larger' | 'largest';
 
@@ -100,41 +101,3 @@ export interface ButtonSlots extends BaseSlots {
 }
 
 export type ButtonSlotProps = SlotProps<ButtonSlots, ButtonProps, React.ButtonHTMLAttributes<HTMLButtonElement>>;
-
-export type ButtonTokenSet = ColorTokenSet & {
-  /* sizing */
-  padding: string;
-  margin: string;
-  height: string;
-  minWidth: string;
-  maxWidth: string;
-  minHeight: string;
-  contentGap: string;
-  iconSize: string;
-  borderRadius: string;
-  borderWidth: string;
-  boxShadow: string;
-  width: string;
-
-  size: {
-    smallest: string;
-    smaller: string;
-    small: string;
-    regular: string;
-    large: string;
-    larger: string;
-    largest: string;
-  };
-
-  transform: string;
-  transition: string;
-
-  fontFamily: string;
-  fontSize: string;
-  fontWeight: string;
-
-  pressed: {
-    transform: string;
-    transition: string;
-  };
-};
