+++ packages/react-button/src/utils/tempTypes.ts
@@ -15,14 +15,6 @@ export interface ComponentProps {
   className?: string;
 }
 
-export type RecursivePartial<T> = {
-  [P in keyof T]?: T[P] extends (infer U)[]
-    ? RecursivePartial<U>[]
-    : T[P] extends object
-    ? RecursivePartial<T[P]>
-    : T[P];
-};
-
 export type ShorthandValue<TProps> = string | boolean | number | null | undefined | TProps | JSX.Element;
 
 export type ComposeRender<TProps, TOptions> = (
