+++ packages/react-theme-provider/etc/react-theme-provider.api.md
@@ -4,60 +4,19 @@
 
 ```ts
 
-import { IPartialTheme } from '@uifabric/styling';
-import { IStyleFunctionOrObject } from '@uifabric/utilities';
+import { ColorTokenSet } from '@fluentui/themes';
+import { PartialTheme } from '@fluentui/themes';
 import * as React from 'react';
-
-// @public
-export type ColorTokens = Partial<{
-    background: string;
-    contentColor: string;
-    subTextColor: string;
-    linkColor: string;
-    iconColor: string;
-    borderColor: string;
-    dividerColor: string;
-    focusColor: string;
-    focusInnerColor: string;
-    opacity: string;
-}>;
+import { Theme } from '@fluentui/themes';
+import { TokenSet } from '@fluentui/themes';
 
 // @public (undocumented)
-export type ColorTokenSet = ColorTokens & ColorTokenStates;
-
-// @public
-export type ColorTokenStates = Partial<{
-    hovered: ColorTokens;
-    pressed: ColorTokens;
-    disabled: ColorTokens;
-    checked: ColorTokens;
-    checkedHovered: ColorTokens;
-    checkedPressed: ColorTokens;
-}>;
-
-// @public (undocumented)
-export type FontTokens = Partial<{
-    fontFamily: string;
-    fontSize: string;
-    fontWeight: string;
-}>;
-
-// @public (undocumented)
-export const getStyleFromPropsAndOptions: <TProps extends StyleProps<import("./types").ColorTokenSet>, TOptions extends StyleOptions<TProps>>(props: TProps, options: TOptions, prefix?: string | undefined) => import("react").CSSProperties;
+export const getStyleFromPropsAndOptions: <TProps extends StyleProps<import("@fluentui/themes").ColorTokenSet>, TOptions extends StyleOptions<TProps>>(props: TProps, options: TOptions, prefix?: string | undefined) => import("react").CSSProperties;
 
 // @public
 export function mergeThemes<TResult = PartialTheme>(...themes: (undefined | PartialTheme | Theme)[]): TResult;
 
-// @public
-export interface PartialTheme extends Omit<Theme, 'tokens'> {
-    // (undocumented)
-    tokens?: RecursivePartial<Tokens>;
-}
-
-// @public
-export type RecursivePartial<T> = {
-    [P in keyof T]?: T[P] extends Array<infer I> ? Array<RecursivePartial<I>> : RecursivePartial<T[P]>;
-};
+export { PartialTheme }
 
 // @public (undocumented)
 export interface StyleOptions<TProps> {
@@ -73,19 +32,7 @@ export interface StyleProps<TTokens extends ColorTokenSet = ColorTokenSet> {
     tokens?: TTokens;
 }
 
-// @public
-export interface Theme extends IPartialTheme {
-    // (undocumented)
-    components?: {
-        [componentName: string]: {
-            styles?: IStyleFunctionOrObject<any, any>;
-        };
-    };
-    // (undocumented)
-    stylesheets?: string[];
-    // (undocumented)
-    tokens?: Tokens;
-}
+export { Theme }
 
 // @public
 export const ThemeProvider: React.ForwardRefExoticComponent<ThemeProviderProps & React.RefAttributes<HTMLDivElement>>;
@@ -95,26 +42,15 @@ export interface ThemeProviderProps extends React.HTMLAttributes<HTMLDivElement>
     theme?: PartialTheme | Theme;
 }
 
-// @public (undocumented)
-export interface Tokens {
-    // (undocumented)
-    [key: string]: TokenSetType;
-    // (undocumented)
-    body: ColorTokenSet & TokenSetType;
-}
-
-// @public
-export type TokenSetType = {
-    [key: string]: TokenSetType | string | number | undefined;
-};
+export { TokenSet }
 
 // @public (undocumented)
-export const tokensToStyleObject: (tokens?: TokenSetType | undefined, prefix?: string | undefined, style?: import("react").CSSProperties | undefined) => import("react").CSSProperties;
+export const tokensToStyleObject: (tokens?: TokenSet | undefined, prefix?: string | undefined, style?: import("react").CSSProperties | undefined) => import("react").CSSProperties;
 
 // @public
 export const useInlineTokens: (draftState: {
     style?: import("react").CSSProperties | undefined;
-    tokens?: TokenSetType | undefined;
+    tokens?: TokenSet | undefined;
 }, prefix: string) => void;
 
 // @public
