--- packages/themes/src/FluentTheme.ts
@@ -1,4 +1,0 @@
-import { Theme } from './types';
-import { createDefaultTheme } from './createDefaultTheme';
-
-export const FluentTheme: Theme = createDefaultTheme();
