+++ packages/themes/src/themes/FluentTheme.ts
@@ -1,0 +1,4 @@
+import { Theme } from '../types';
+import { createDefaultTheme } from '../createDefaultTheme';
+
+export const FluentTheme: Theme = createDefaultTheme();
